+++ websocketpp/endpoint.hpp
@@ -32,7 +32,6 @@
 #include <websocketpp/logger/levels.hpp>
 #include <websocketpp/version.hpp>
 
-#include <iostream>
 #include <set>
 
 namespace websocketpp {
@@ -88,8 +87,8 @@ public:
     typedef lib::shared_ptr<connection_weak_ptr> hdl_type;
 
     explicit endpoint(bool p_is_server)
-      : m_alog(config::alog_level, &std::cout)
-      , m_elog(config::elog_level, &std::cerr)
+      : m_alog(config::alog_level, log::channel_type_hint::access)
+      , m_elog(config::elog_level, log::channel_type_hint::error)
       , m_user_agent(::websocketpp::user_agent)
       , m_open_handshake_timeout_dur(config::timeout_open_handshake)
       , m_close_handshake_timeout_dur(config::timeout_close_handshake)